raspberry-pi/requirements.txt
opencv-python>=4.8.0
requests>=2.31.0
pyserial>=3.5
numpy>=1.24.0

# Optionnel (pour plus tard)
# face-recognition>=1.3.0
# easyocr>=1.7.0
# torch>=2.0.0